+++ cmd/core/ports/find-user.service.go
@@ -1,0 +1,11 @@
+package ports
+
+import (
+	entity "go-api/cmd/core/entities/user"
+)
+
+type (
+	GetUserService interface {
+		GetUser(id string) (*entity.User, error)
+	}
+)