+++ cmd/core/use-case/create-user-producer/interface.go
@@ -1,0 +1,11 @@
+package createuserproducerusecase
+
+import (
+	port "go-api/cmd/core/ports"
+)
+
+type (
+	CreateUserUseCase interface {
+		CreateUser(dto port.CreateDto) error
+	}
+)