+++ cmd/core/use-case/list-user/main.go
@@ -1,0 +1,17 @@
+package listusersusecase
+
+import (
+	"go-api/cmd/core/ports"
+)
+
+type listUsersUseCase struct {
+	IntegrationJsonPlaceHolder ports.JsonPlaceholderIntegration
+	UsersCache                 ports.UsersCache
+}
+
+func New(ji ports.JsonPlaceholderIntegration, uc ports.UsersCache) ListUsersUseCase {
+	return &listUsersUseCase{
+		IntegrationJsonPlaceHolder: ji,
+		UsersCache:                 uc,
+	}
+}