+++ cmd/infra/adapters/mysql/migrations.go
@@ -1,0 +1,11 @@
+package database
+
+import (
+	repositoryUser "go-api/cmd/infra/repositories/sql/user"
+
+	"github.com/jinzhu/gorm"
+)
+
+func LoadMigrationByRepositores(db *gorm.DB) {
+	repositoryUser.InitMigrate(db)
+}