+++ main.go
@@ -1,6 +1,6 @@
 package main
 
-import app "go-api/src/main"
+import app "go-api/cmd/main"
 
 func main() {
 	app.Start()