+++ src/infra/integrations/amqp/client/rabbitmq-client-interface.go
@@ -1,7 +1,0 @@
-package amqpclient
-
-import typesclient "go-api/src/infra/integrations/amqp/client/types"
-
-type AmqpClient interface {
-	Publish(b []byte, c typesclient.ConfigAmqpClient) error
-}